decorator.lua: create helper submodule allowing for better documentation
Modules required by doc cannot be documented with doc itself.  Now doc
requires the submodule and the top level decorator module is free to be
documented normally.
