+++ luamodules/decorator/intern/decorator_test.go
@@ -1,20 +1,14 @@
-package decorator
+package intern
 
 import (
 	"testing"
 
-	"github.com/bmatsuo/lark/luamodules/decorator/intern"
-	"github.com/bmatsuo/lark/luamodules/doc"
 	"github.com/bmatsuo/lark/luatest"
 )
 
 var testModule = &luatest.Module{
 	Module:     Module,
 	TestScript: "decorator_test.lua",
-	PreloadDeps: []*luatest.Module{
-		{Module: doc.Module},
-		{Module: intern.Module},
-	},
 }
 
 func TestModule(t *testing.T) {
