+++ luamodules/doc/doc_test.go
@@ -3,7 +3,7 @@ package doc
 import (
 	"testing"
 
-	"github.com/bmatsuo/lark/luamodules/decorator"
+	"github.com/bmatsuo/lark/luamodules/decorator/intern"
 	"github.com/bmatsuo/lark/luatest"
 )
 
@@ -11,7 +11,7 @@ var testModule = &luatest.Module{
 	Module:     Module,
 	TestScript: "doc_test.lua",
 	PreloadDeps: []*luatest.Module{
-		{Module: decorator.Module},
+		{Module: intern.Module},
 	},
 }
