+++ luamodules/lark/task/task_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/bmatsuo/lark/luamodules/decorator"
+	"github.com/bmatsuo/lark/luamodules/decorator/intern"
 	"github.com/bmatsuo/lark/luamodules/doc"
 	"github.com/bmatsuo/lark/luatest"
 )
@@ -14,6 +15,7 @@ var testModule = &luatest.Module{
 	PreloadDeps: []*luatest.Module{
 		{Module: doc.Module},
 		{Module: decorator.Module},
+		{Module: intern.Module},
 	},
 }
