+++ luamodules/path/path_test.go
@@ -3,7 +3,7 @@ package path
 import (
 	"testing"
 
-	"github.com/bmatsuo/lark/luamodules/decorator"
+	"github.com/bmatsuo/lark/luamodules/decorator/intern"
 	"github.com/bmatsuo/lark/luamodules/doc"
 	"github.com/bmatsuo/lark/luatest"
 )
@@ -13,7 +13,7 @@ var testModule = &luatest.Module{
 	TestScript: "path_test.lua",
 	PreloadDeps: []*luatest.Module{
 		{Module: doc.Module},
-		{Module: decorator.Module},
+		{Module: intern.Module},
 	},
 }
